ajax insertion for items

+++ NMS-Inventory System/app/Http/Controllers/ItemsController.php
@@ -14,6 +14,7 @@ use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidateRequests;
 use Illuminate\Foundation\Validation\AuthorizesRequests;
 use Illuminate\Support\Facades\Redirect;
+use Validator, Input; 
 
 class ItemsController extends Controller
 {
@@ -39,6 +40,7 @@ class ItemsController extends Controller
         else{
             return view('pages/items_page');
         }
+       // return Datatables::of($students)->make(true);
     }
 
 
@@ -171,33 +173,33 @@ class ItemsController extends Controller
         return back()->with($notification);
     }
 
+    function insert(Request $req)
+    {
+            if($req->get('button_action') == "insert")
+            {
+                $itemname = $req->input('itemname');
+                $itemdesc = $req->input('itemdesc');
+                $price = $req->input('price');
+                $quantity = $req->input('quantity');
+                $catid = $req->input('catid');
+                $item =  array('itemname'=>$itemname,'itemdesc'=>$itemdesc,'price'=>$price,'quantity'=>$quantity,'catid'=>$catid,'created_at'=>NOW(),'updated_at'=>NULL,'deleted_at'=>NULL);
 
-    function insert(Request $req){
-        $itemname = $req->input('itemname');
-        $itemdesc = $req->input('itemdesc');
-        $price = $req->input('price');
-        $quantity = $req->input('quantity');
-        $catid = $req->input('catid');
-        $data = array('itemname'=>$itemname,'itemdesc'=>$itemdesc,'price'=>$price,'quantity'=>$quantity,'catid'=>$catid,'created_at'=>NOW(),'updated_at'=>NULL,'deleted_at'=>NULL);
-
-
-        if (DB::table('items')->where('itemname', '=', $itemname)->exists()) {
-            DB::table('items')->where('itemname', '=', $itemname)->delete();
-            DB::table('items')->insert($data);
-            $notification = array(
-                'message'=> 'A New Item is Inserted!',
-                'alert-type' => 'success'
-            );
-        }else{
-            DB::table('items')->insert($data);
-            $notification = array(
-                'message'=> 'A New Item is Inserted!',
-                'alert-type' => 'success'
-            );
-        }
-        return back()->with($notification);
-
-
+                if (DB::table('items')->where('itemname', '=', $itemname)->exists()) {
+                    DB::table('items')->where('itemname', '=', $itemname)->delete();
+                    DB::table('items')->insert($item);
+                    $notification = array(
+                        'message'=> 'A New Item is Inserted!',
+                        'alert-type' => 'success'
+                    );
+                }else{
+                    DB::table('items')->insert($item);
+                    $notification = array(
+                        'message'=> 'A New Item is Inserted!',
+                        'alert-type' => 'success'
+                    );
+                }
+            }
+            return back()->with($notification);
     }
 }
